# job-rotation-app-backend/openApiSpec.yml
openapi: '3.0.3'
info:
  title: Job Rotation Application APIs
  version: '1.0'
servers:
  - url: https://api.server.test/v1
components:
  schemas:
    ListAllRolesResponse:
      description: Response containing the summary of available roles to be displayed on the employee landing page
      type: object
      required:
        - roleId
        - roleName
        - gradeRequired
        - department
      properties:
        roleId:
          type: string
        roleName:
          type: string
        gradeRequired:
          type: string
        department:
          type: string
    GetRoleDetailsResponse:
      description: Response containing the detailed information of a specific role
      type: object
      required:
        - roleId
        - roleName
        - gradeRequired
        - department
        - staffingManagerEmailAddress
        - duration
        - jobDescription
      properties:
        roleId:
          type: string
        roleName:
          type: string
        gradeRequired:
          type: string
        staffingManagerEmailAddress:
          type: string
        department:
          type: string
        duration:
          type: string
        jobDescription:
          type: string
paths:
  /employee/list-available-roles:
    get:
      tags:
        - Employee Actions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAllRolesResponse'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /employee/list-available-roles/{grade}:
    get:
      tags:
        - Employee Actions
      parameters:
        - name: grade
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAllRolesResponse'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /employee/list-available-roles/{department}:
    get:
      tags:
        - Employee Actions
      parameters:
        - name: department
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAllRolesResponse'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /employee/list-available-roles/{duration}:
    get:
      tags:
        - Employee Actions
      parameters:
        - name: duration
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAllRolesResponse'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
...
  /employee/available-roles/{roleId}:
    get:
      tags:
        - Role Information
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRoleDetailsResponse'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
...